updated ft_itoa.c

+++ Libft/ft_itoa.c
@@ -6,13 +6,13 @@
 /*   By: jngew <jngew@student.42singapore.sg>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/23 23:56:27 by jngew             #+#    #+#             */
-/*   Updated: 2024/05/24 17:33:48 by jngew            ###   ########.fr       */
+/*   Updated: 2024/05/24 17:38:29 by jngew            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static int	get_dcount(int n)
+static int	get_dcount(long n)
 {
 	int	count;
